EVM, VM: moved supported HF check to EVM to cover both EVM and VM
